+++ resources/views/auth/passwords/email.blade.php
@@ -9,7 +9,7 @@
     <div class="container min-h-screen flex items-center">
 		<div class="row w-full justify-center my-32">
 			<div class="xs-column-12 s-column-10 m-column-6 l-column-4 relative z-20">
-                    
+
                 <h1 class="mb-8">Je wachtwoord vergeten?</h1>
                 <p>Geef je e-mailadres in om je wachtwoord te opnieuw in te stellen.</p>
 
@@ -28,9 +28,9 @@
                         <div class="stack">
                             <input type="email" class="inset-s" name="email" placeholder="E-mail" id="identity" value="{{ old('email') }}">
                         </div>
-                        
+
                         <button type="submit" class="btn btn-primary mr-4">Reset mijn wachtwoord</button>
-                        <a href="{{ route('chief.back.login') }}">Keer terug naar de login</a>
+                        <a href="{{ route('chief.back.login') }}">Terug naar login</a>
 
                     </form>